pyproject.toml
[project]
name = "rtsp-viewer"
version = "0.1.0"
description = "A world-class RTSP stream viewer and recorder using OpenCV and ffmpeg"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "opencv-python>=4.8.0",
    "numpy>=1.24.0",
    "pyyaml>=6.0",
    "pillow>=10.0.0",
]

[project.scripts]
rtsp-viewer = "main:main"

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]
